+++ src/App.vue
@@ -1,42 +1,26 @@
 <script setup lang="ts">
-import { ref } from "vue";
-import { jyongDziwbyoKriemsakkho } from "./stores/dziwbyoKriemsak";
 import {
   jyongDziwqrimKriemsakkho,
   type Dziwqrim,
 } from "./stores/dziwqrimKriemsak";
 import { jyongDziwdeukho } from "./stores/dziwdeu";
-import shiwxryn from "./sryokio/shiwxryn.json";
-import shiengmu from "./sryokio/shiengmu.json";
+import KriemsakPenlan from "./components/KriemsakPenlan.vue";
+import { jyongXeithongkho } from "./stores/xeithong";
 
-type Thryuthei = boolean | null;
-type PenlanPieuchiem = string | null;
-
-const dziwbyoKriemsakkho = jyongDziwbyoKriemsakkho();
 const dziwqrimKriemsakkho = jyongDziwqrimKriemsakkho();
 const dziwdeu = jyongDziwdeukho();
-
-const thryuthei = ref(<Thryuthei>null);
-const penlanPieuchiem = ref(<PenlanPieuchiem>null);
+const xeithongkho = jyongXeithongkho();
 
 const qanZyepheng = (tuiziang: Dziwqrim) => {
   dziwqrimKriemsakkho.triwSrioNipZyepheng(tuiziang.zyepheng);
-  penlanPieuchiem.value = "qrit";
+  xeithongkho.kriemsakPenlanPieuchiem = "qrit";
 };
 
 const qanShiwxryn = (tuiziang: Dziwqrim) => {
   dziwqrimKriemsakkho.triwSrioSyenShiwxryn(tuiziang.shiwxryn);
-  penlanPieuchiem.value = "qrit";
+  xeithongkho.kriemsakPenlanPieuchiem = "qrit";
 };
 
-const qanKriemsakKetkua = (tuiziang: string) => {
-  dziwdeu.triwSrioSyenDziwdu(tuiziang);
-};
-
-const xuanThryuthei = () => {
-  thryuthei.value = !thryuthei.value;
-};
-const qhruaSikrokXaumra = (xaumraco: string[]) => xaumraco.join("、");
 const twkDryenthungQrimlyi = (dziwqrim: Dziwqrim) => {
   return [
     dziwqrim.shiep,
@@ -79,105 +63,14 @@ const twkQhZQh = (dziwqrim: Dziwqrim) => {
     <v-system-bar> </v-system-bar>
 
     <v-app-bar>
-      <v-app-bar-nav-icon @click="xuanThryuthei()"></v-app-bar-nav-icon>
+      <v-app-bar-nav-icon
+        @click="xeithongkho.xuanKriemsakPenlanThryuthei()"
+      ></v-app-bar-nav-icon>
       <v-toolbar-title>網絡本《漢字文聲義》</v-toolbar-title>
       <v-spacer />
     </v-app-bar>
 
-    <v-navigation-drawer v-model="thryuthei" bottom>
-      <v-tabs v-model="penlanPieuchiem" fixed-tabs>
-        <v-tab value="krap">字符</v-tab>
-        <v-tab value="qrit">字音</v-tab>
-      </v-tabs>
-      <v-window v-model="penlanPieuchiem">
-        <v-window-item value="krap">
-          <v-sheet class="pa-4">
-            <v-text-field
-              v-model="dziwbyoKriemsakkho.srioNipQhandziwthyen"
-              label="漢字"
-              placeholder="天地玄黄宇宙洪荒"
-              clearable
-            />
-            <v-btn
-              block
-              rounded
-              @click="dziwbyoKriemsakkho.triwKriemsakKetkua()"
-            >
-              檢索
-            </v-btn>
-          </v-sheet>
-
-          <v-divider></v-divider>
-
-          <v-list>
-            <v-list-item
-              v-for="tuiziang in dziwbyoKriemsakkho.twkKriemsakKetkua"
-              :key="tuiziang.pieutiwbyo"
-              @click="qanKriemsakKetkua(tuiziang.dziw)"
-              link
-            >
-              <v-list-item-title>{{ tuiziang.dziw }}</v-list-item-title>
-              {{ qhruaSikrokXaumra(tuiziang.sikrok_xaumra) }}
-            </v-list-item>
-          </v-list>
-        </v-window-item>
-        <v-window-item value="qrit">
-          <v-sheet class="pa-4">
-            <v-col>
-              <v-text-field
-                v-model="dziwqrimKriemsakkho.srioNipZyepheng"
-                label="隋拼"
-                placeholder="Thendʼî xuenxuaŋ"
-                density="compact"
-                variant="underlined"
-                clearable
-              />
-              <v-select
-                label="平水韵"
-                v-model="dziwqrimKriemsakkho.srioSyenShiwxryn"
-                :items="shiwxryn"
-                item-title="xryn_driang"
-                item-value="xryn_tuan"
-                variant="underlined"
-                density="compact"
-                clearable
-              />
-              <v-select
-                label="聲母"
-                v-model="dziwqrimKriemsakkho.srioSyenShiengmu"
-                :items="shiengmu"
-                item-title="shiengmu_ziangdzieng"
-                item-value="shiengmu"
-                variant="underlined"
-                density="compact"
-                clearable
-              />
-            </v-col>
-            <v-btn
-              block
-              rounded
-              @click="dziwqrimKriemsakkho.triwKriemsakKetkua"
-            >
-              檢索
-            </v-btn>
-          </v-sheet>
-
-          <v-divider></v-divider>
-
-          <v-list>
-            <v-list-item
-              v-for="tuiziang in dziwqrimKriemsakkho.twkKriemsakKetkua"
-              :key="tuiziang.pieutiwbyo"
-              @click="qanKriemsakKetkua(tuiziang.dziw)"
-              link
-            >
-              <v-list-item-title>{{ tuiziang.dziw }}</v-list-item-title>
-              {{ tuiziang.zyepheng }}
-            </v-list-item>
-          </v-list>
-        </v-window-item>
-      </v-window>
-    </v-navigation-drawer>
+    <KriemsakPenlan />
 
     <v-main>
       <v-container fluid>
