+++ src/stores/xeithong.ts
@@ -5,8 +5,16 @@ type Pieuchiem = string | null;
 type Thryuthei = boolean | null;
 
 export const jyongXeithongkho = defineStore("xeithong", () => {
-  const KriemsakPenlanPieuchiem = ref(<Pieuchiem>null);
-  const KriemsakPenlanThryuthei = ref(<Thryuthei>null);
+  const kriemsakPenlanPieuchiem = ref(<Pieuchiem>null);
+  const kriemsakPenlanThryuthei = ref(<Thryuthei>null);
 
-  return { KriemsakPenlanPieuchiem, KriemsakPenlanThryuthei };
+  const xuanKriemsakPenlanThryuthei = () => {
+    kriemsakPenlanThryuthei.value = !kriemsakPenlanThryuthei.value;
+  };
+
+  return {
+    kriemsakPenlanPieuchiem,
+    kriemsakPenlanThryuthei,
+    xuanKriemsakPenlanThryuthei,
+  };
 });
